Fixed register route to make it comply with Model, View, Controller infastrucutre
