+++ backend/routes/users.js
@@ -1,6 +1,6 @@
 const { application } = require('express');
 const express = require('express');
-const users = require('../models/users.js');
+const userController = require('../controllers/users.js');
 
 
 /**
@@ -19,14 +19,17 @@ router.post('/register', async (req, res, next) => {
     //    const result = await req.models.user.createNewUser(body.email, body.password);
        // const result = await req.models.createNewUser(req.body.firstName, req.body.lastName, req.body.email, req.body.password
       //calls the createNewUser function in the users.js file of the models folder and return the result
-        const result = await users.createNewUser(body.firstName, body.lastName, body.email, body.password);
+        const result = await userController.createNewUser(body.firstName, body.lastName, body.email, body.password);
         console.log("Result of createNewUser: ", result);
         if(result.error === "Invalid email"){
             console.log("Invalid email");
             res.status(400).json({message: "Invalid email"});
-        } else if(result.error === "User already exists"){
+        } else if(result.error === "User already exists") {
             console.log("User already exists");
             res.status(400).json({message: "User already exists"});
+        } else if(result.error === "Incomplete input") {
+            console.log("Incomplete input");
+            res.status(400).json({message:"Incomplete input"});
         } else {
             console.log("User created");
             res.status(201).json(result);
